refactor package structutil to structs
