+++ structs/struct_internal.go
@@ -1,4 +1,4 @@
-package structutil
+package structs
 
 import "fmt"
